+++ MQTTPacket/src/V3/MQTTConnect.h
@@ -21,7 +21,7 @@
 #define MQTTCONNECT_H_
 
 #include <stdint.h>
-#include "MQTTConnectInternal.h"
+#include "../MQTTConnectInternal.h"
 
 enum MQTTConnackReturnCodes
 {
@@ -79,15 +79,15 @@ typedef struct
 #define MQTTPacket_connectData_initializer { {'M', 'Q', 'T', 'C'}, 4, {NULL, {0, NULL}}, 60, 1, 0, \
 		MQTTPacket_willOptions_initializer, {NULL, {0, NULL}}, {NULL, {0, NULL}} }
 
-DLLExport int MQTTSerialize_connect(unsigned char* buf, int32_t buflen, MQTTPacket_connectData* options);
+DLLExport int MQTTSerialize_connect(unsigned char* buf, size_t buflen, MQTTPacket_connectData* options);
 DLLExport int MQTTDeserialize_connect(MQTTPacket_connectData* data, unsigned char* buf, int32_t len);
 
-DLLExport int MQTTSerialize_connack(unsigned char* buf, int32_t buflen, unsigned char connack_rc, unsigned char sessionPresent);
-DLLExport int MQTTDeserialize_connack(unsigned char* sessionPresent, unsigned char* connack_rc, unsigned char* buf, int32_t buflen);
+DLLExport int MQTTSerialize_connack(unsigned char* buf, size_t buflen, unsigned char connack_rc, unsigned char sessionPresent);
+DLLExport int MQTTDeserialize_connack(unsigned char* sessionPresent, unsigned char* connack_rc, unsigned char* buf, size_t buflen);
 
-DLLExport int MQTTSerialize_disconnect(unsigned char* buf, int32_t buflen);
-DLLExport int MQTTDeserialize_disconnect(unsigned char* buf, int32_t buflen);
+DLLExport int MQTTSerialize_disconnect(unsigned char* buf, size_t buflen);
+DLLExport int MQTTDeserialize_disconnect(unsigned char* buf, size_t buflen);
 
-DLLExport int MQTTSerialize_pingreq(unsigned char* buf, int32_t buflen);
+DLLExport int MQTTSerialize_pingreq(unsigned char* buf, size_t buflen);
 
 #endif /* MQTTCONNECT_H_ */
